backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
langchain
langchain-openai
chromadb
python-dotenv
pydantic
pypdf
requests
python-multipart
tiktoken
beautifulsoup4
